services/extractor_service/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
yt-dlp==2025.04.30
faster-whisper==1.1.1 
protobuf==5.29.4
PyYAML==6.0.2
